code/src/python/regulatory-ai-backend/app/requirements.txt.txt
flask==2.3.2
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
openai==0.27.8
python-dotenv==1.0.0
gunicorn==20.1.0
pydantic==1.10.7